# ObjectGraphUtils traversal optimizations