--- a/src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/util/FrameAccess.java
+++ b/src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/util/FrameAccess.java
@@ -479,6 +479,14 @@ public final class FrameAccess {
         CompilerDirectives.bailout("Finding materializable frames should never be part of compiled code as it triggers deopts");
         LogUtils.ITERATE_FRAMES.fine("Iterating frames to find a marker...");
         final Frame frame = Truffle.getRuntime().iterateFrames(frameInstance -> {
+            if (frameInstance.getCallTarget() instanceof final RootCallTarget rct && rct.getRootNode() instanceof final ResumeContextRootNode rcrn) {
+                /*
+                 * Reached end of Smalltalk activations on Truffle frames.
+                 */
+                final ContextObject context = rcrn.getActiveContext();
+                assert context.getFrameMarker() == frameMarker : "Failed to find frameMarker in ResumeContextRootNode";
+                return context.getTruffleFrame();
+            }
             final Frame current = frameInstance.getFrame(FrameInstance.FrameAccess.READ_ONLY);
             if (!isTruffleSqueakFrame(current)) {
                 return null;
@@ -490,22 +498,6 @@ public final class FrameAccess {
             return null;
         });
         if (frame == null) {
-
-            // Try to find the resume context, if any.
-            final ContextObject resumeContextObject = Truffle.getRuntime().iterateFrames(frameInstance -> {
-                if (frameInstance.getCallTarget() instanceof final RootCallTarget rct && rct.getRootNode() instanceof final ResumeContextRootNode rcrn) {
-                    /*
-                     * Reached end of Smalltalk activations on Truffle frames. From here, tracing
-                     * should continue to walk senders via ContextObjects.
-                     */
-                    return rcrn.getActiveContext(); // break
-                }
-                return null;
-            });
-            if (resumeContextObject != null && resumeContextObject.getFrameMarker() == frameMarker) {
-                return resumeContextObject.getTruffleFrame();
-            }
-
             throw SqueakException.create("Could not find frame for:", frameMarker);
         } else {
             return frame.materialize();