Changed ObjectTracer to mark unmarked objects when adding them to the workStack. This avoids an extra check in all of the traversals and eliminates the potential examination of an object multiple times.

Since the workStack is a stack, reordered the initialization to put the specialObjectsArray (and its contents) at the top of the stack and changed tracePointers to add the class after the inner pointers so that the class would be traversed before the other pointers.

Rearranged pointersBecomeOneWay and allInstancesOf so that up to 4 tasks can work independently on subgraphs of the entire object graph.