@fniephaus 2025-05-20
Nice! Using the following as a benchmark:

```
FileStream stdout nextPutAll: (ClassBuilderTest buildSuite run asString); cr; flush.
FileStream stdout nextPutAll: (ClassBuilderTest buildSuite run asString); cr; flush.
Smalltalk snapshot: false andQuit: true.
```

I get, before:

```
$ mx build && $GRAALVM_HOME/bin/trufflesqueak --smalltalk.resource-summary --experimental-options images/test-64bit.1.image /Users/fniephaus/dev/graalvm/trufflesqueak/test.st
...
[trufflesqueak] Running test-64bit.1.image on Interpreted...
[trufflesqueak] Image loaded in 325ms.
19 run in 0:00:00:30.278, 19 passes, 0 expected failures, 0 failures, 0 errors, 0 unexpected passes
19 run in 0:00:00:29.712, 19 passes, 0 expected failures, 0 failures, 0 errors, 0 unexpected passes
[trufflesqueak] # Resource Summary
[trufflesqueak] - Total process time: 62.006s | CPU load: 1.11
[trufflesqueak] 
[trufflesqueak] ## GC Statistics
[trufflesqueak] -     0.0810s ( 0.13% of total time) in  120 GCs of G1 Young Generation
[trufflesqueak] -     0.0010s ( 0.00% of total time) in    2 GCs of G1 Concurrent GC
[trufflesqueak] -     0.0570s ( 0.09% of total time) in    2 GCs of G1 Old Generation
[trufflesqueak] -     0.1390s ( 0.22% of total time) in  124 GCs in total
[trufflesqueak] 
[trufflesqueak] ## Object Graph Operations Statistics
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'allInstances'
[trufflesqueak] -     0.8380s ( 1.35% of total time) for   31 'allInstancesOf'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'someInstanceOf'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'nextObject'
[trufflesqueak] -    31.3470s (50.55% of total time) for 1257 'pointersBecomeOneWay'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'checkEphemerons'
```

After:
```
$ mx build && $GRAALVM_HOME/bin/trufflesqueak --smalltalk.resource-summary --experimental-options images/test-64bit.1.image /Users/fniephaus/dev/graalvm/trufflesqueak/test.st
...
[trufflesqueak] Running test-64bit.1.image on Interpreted...
[trufflesqueak] Image loaded in 318ms.
19 run in 0:00:00:18.281, 19 passes, 0 expected failures, 0 failures, 0 errors, 0 unexpected passes
19 run in 0:00:00:18.252, 19 passes, 0 expected failures, 0 failures, 0 errors, 0 unexpected passes
[trufflesqueak] # Resource Summary
[trufflesqueak] - Total process time: 38.383s | CPU load: 1.72
[trufflesqueak] 
[trufflesqueak] ## GC Statistics
[trufflesqueak] -     0.0870s ( 0.23% of total time) in  130 GCs of G1 Young Generation
[trufflesqueak] -     0.0010s ( 0.00% of total time) in    2 GCs of G1 Concurrent GC
[trufflesqueak] -     0.0570s ( 0.15% of total time) in    2 GCs of G1 Old Generation
[trufflesqueak] -     0.1450s ( 0.38% of total time) in  134 GCs in total
[trufflesqueak] 
[trufflesqueak] ## Object Graph Operations Statistics
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'allInstances'
[trufflesqueak] -     0.3710s ( 0.97% of total time) for   31 'allInstancesOf'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'someInstanceOf'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'nextObject'
[trufflesqueak] -    12.1300s (31.60% of total time) for 1175 'pointersBecomeOneWay'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'checkEphemerons'
```

38.383s are ~61.9% of 62.006s total process time. Well done!