@Sabine-Schnabeltier 2025-05-21
The following test code is showing that the parallel version is incorrectly counting **allInstances**:

```
Object >> test
	| theClass array failing |
	
	theClass := Dictionary.

	array := theClass allInstances.

	Transcript cr; show: theClass allInstances size printString.
	
	failing := OrderedCollection new.

	1 to: 1000 do: [ :i | | all |
		all := theClass allInstances.
		all size = array size ifFalse: [
			failing addLast: all.
			Transcript cr; show: 'FAIL[', i printString, ']: ', all size printString]].
	
	^ { array . failing }

"Object test."
```

It appears to be double traversing some subgraphs due to a race condition in `AbstractSqueakObjectWithClassAndHash.tryToMark()`.