@Sabine-Schnabeltier 2025-05-22
Adding `synchronized` to `AbstractSqueakObjectWithClassAndHash.tryToMark()` fixes the double traversal problem but the overhead results in some backtracking on the earlier speed improvement.