@Sabine-Schnabeltier 2025-05-23
The double traversal problem might also leave the object graph in an inconsistent marking state.

Reimplemented the mark bit routines using lock-free atomic operations rather than using `synchronized` and got back the original speed improvement (perhaps even more?).