@fniephaus 2025-05-25
The latest results are in:

**[Before (on ubuntu-arm)](https://github.com/hpi-swa/trufflesqueak/actions/runs/15158541125/job/42619020913#step:8:11487):**
```
[trufflesqueak] ## Object Graph Operations Statistics
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'allInstances'
[trufflesqueak] -    15.6430s ( 2.30% of total time) for  201 'allInstancesOf'
[trufflesqueak] -     0.0590s ( 0.01% of total time) for    2 'someInstanceOf'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'nextObject'
[trufflesqueak] -   301.7440s (44.30% of total time) for 3640 'pointersBecomeOneWay'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'checkEphemerons'
de.hpi.swa.trufflesqueak.test.SqueakSUnitTest finished 658456.2 ms, [disk (free/total): 37.1GB/72.5GB]
Time: 679.94
```

**[After (on ubuntu-arm)](https://github.com/hpi-swa/trufflesqueak/actions/runs/15236224100/job/42850566622#step:8:11342):**
```
[trufflesqueak] ## Object Graph Operations Statistics
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'allInstances'
[trufflesqueak] -     9.9380s ( 1.88% of total time) for  197 'allInstancesOf'
[trufflesqueak] -     0.0690s ( 0.01% of total time) for    2 'someInstanceOf'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'nextObject'
[trufflesqueak] -   158.0030s (29.91% of total time) for 3719 'pointersBecomeOneWay'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'checkEphemerons'
de.hpi.swa.trufflesqueak.test.SqueakSUnitTest finished 506776.3 ms, [disk (free/total): 37.1GB/72.5GB]
Time: 526.985
```

So on the new ubuntu-arm runners, the tests ran in ~77.50% of the time, with `pointersBecomeOneWay` running in only ~52,4% of the time it used to. We still need to take these numbers with a grain of salt because the runners don't provide consistent performance and are thus bad for benchmarking, but the improvements are quite significant. Great!

However, the coverage gate actually regressed a bit. I suspect this is due to a slower runner.

**[Before (on ubuntu with coverage)](https://github.com/hpi-swa/trufflesqueak/actions/runs/15158541125/job/42619020902#step:9:9507):**
```
[trufflesqueak] ## Object Graph Operations Statistics
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'allInstances'
[trufflesqueak] -    14.9520s ( 1.49% of total time) for  193 'allInstancesOf'
[trufflesqueak] -     0.0550s ( 0.01% of total time) for    2 'someInstanceOf'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'nextObject'
[trufflesqueak] -   359.5540s (35.88% of total time) for 4442 'pointersBecomeOneWay'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'checkEphemerons'
de.hpi.swa.trufflesqueak.test.SqueakSUnitTest finished 985686.9 ms, [disk (free/total): 13.0GB/72.5GB]
Time: 1,001.142
```

**[After (on ubuntu with coverage)](https://github.com/hpi-swa/trufflesqueak/actions/runs/15236224100/job/42850566625#step:9:9715):**
```
[trufflesqueak] ## Object Graph Operations Statistics
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'allInstances'
[trufflesqueak] -    19.1830s ( 1.84% of total time) for  209 'allInstancesOf'
[trufflesqueak] -     0.0730s ( 0.01% of total time) for    2 'someInstanceOf'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'nextObject'
[trufflesqueak] -   407.5480s (39.17% of total time) for 4434 'pointersBecomeOneWay'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'checkEphemerons'
de.hpi.swa.trufflesqueak.test.SqueakSUnitTest finished 1024217.2 ms, [disk (free/total): 13.0GB/72.5GB]
Time: 1,039.423
```

On macos-14, on the other hand, the object graph operation statistics have improved, even though the overall test execution didn't change much. I suspect this is also due to a slower runner.

**[Before (on macos-14)](https://github.com/hpi-swa/trufflesqueak/actions/runs/15158541125/job/42619020896#step:8:12202):**
```
[trufflesqueak] ## Object Graph Operations Statistics
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'allInstances'
[trufflesqueak] -    12.9660s ( 2.00% of total time) for  233 'allInstancesOf'
[trufflesqueak] -     0.0610s ( 0.01% of total time) for    2 'someInstanceOf'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'nextObject'
[trufflesqueak] -   234.9600s (36.33% of total time) for 3784 'pointersBecomeOneWay'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'checkEphemerons'
de.hpi.swa.trufflesqueak.test.SqueakSUnitTest finished 626166.7 ms, [disk (free/total): 38.5GB/319.5GB]
Time: 646.127
```

**[After (on macos-14)](https://github.com/hpi-swa/trufflesqueak/actions/runs/15236224100/job/42850566635#step:8:12430):**
```
[trufflesqueak] ## Object Graph Operations Statistics
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'allInstances'
[trufflesqueak] -     9.5210s ( 1.48% of total time) for  241 'allInstancesOf'
[trufflesqueak] -     0.0800s ( 0.01% of total time) for    2 'someInstanceOf'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'nextObject'
[trufflesqueak] -   175.7070s (27.36% of total time) for 4070 'pointersBecomeOneWay'
[trufflesqueak] -     0.0000s ( 0.00% of total time) for    0 'checkEphemerons'
de.hpi.swa.trufflesqueak.test.SqueakSUnitTest finished 618669.4 ms, [disk (free/total): 38.5GB/319.5GB]
Time: 640.562
```